requirements.txt
wget>=3.2
tensorflow>=2.1.1
scikit-learn>=0.18
numpy>=1.9.2
pandas>=0.19
seaborn>=0.11.1
matplotlib>=1.4.3
dvc
google-api-python-client
gcsfs
google-cloud
google-cloud-storage